separated controller into individual routes.
